Move mask_value function to integration_schema.py for better organization and maintainability
